opencti-trend-connector/requirements.txt
pycti==6.4.10
requests
python-dotenv
